requirements.txt
typer>=0.12
rich>=13.0
jinja2>=3.1
PyYAML>=6.0
requests>=2.31
dnspython>=2.4
markdown>=3.5

